Move comfyWorkflow to schemas/
