+++ src/components/sidebar/tabs/QueueSidebarTab.vue
@@ -104,6 +104,7 @@ import { useI18n } from 'vue-i18n'
 
 import NoResultsPlaceholder from '@/components/common/NoResultsPlaceholder.vue'
 import VirtualGrid from '@/components/common/VirtualGrid.vue'
+import { ComfyNode } from '@/schemas/comfyWorkflow'
 import { api } from '@/scripts/api'
 import { app } from '@/scripts/app'
 import { useLitegraphService } from '@/services/litegraphService'
@@ -114,7 +115,6 @@ import {
   useQueueStore
 } from '@/stores/queueStore'
 import { useSettingStore } from '@/stores/settingStore'
-import { ComfyNode } from '@/types/comfyWorkflow'
 
 import SidebarTabTemplate from './SidebarTabTemplate.vue'
 import ResultGallery from './queue/ResultGallery.vue'
