+++ src/components/sidebar/tabs/queue/TaskItem.vue
@@ -73,10 +73,10 @@ import Button from 'primevue/button'
 import Tag from 'primevue/tag'
 import { computed, onMounted, onUnmounted, ref } from 'vue'
 
+import { ComfyNode } from '@/schemas/comfyWorkflow'
 import { api } from '@/scripts/api'
 import { useLitegraphService } from '@/services/litegraphService'
 import { TaskItemDisplayStatus, type TaskItemImpl } from '@/stores/queueStore'
-import { ComfyNode } from '@/types/comfyWorkflow'
 
 import ResultItem from './ResultItem.vue'
