+++ src/composables/usePaste.ts
@@ -2,10 +2,10 @@ import { LiteGraph } from '@comfyorg/litegraph'
 import type { LGraphNode } from '@comfyorg/litegraph'
 import { useEventListener } from '@vueuse/core'
 
+import { ComfyWorkflowJSON } from '@/schemas/comfyWorkflow'
 import { app } from '@/scripts/app'
 import { useCanvasStore } from '@/stores/graphStore'
 import { useWorkspaceStore } from '@/stores/workspaceStore'
-import { ComfyWorkflowJSON } from '@/types/comfyWorkflow'
 import { isAudioNode, isImageNode, isVideoNode } from '@/utils/litegraphUtil'
 
 /**
