+++ src/extensions/core/groupNode.ts
@@ -3,12 +3,16 @@ import { type LGraph, LGraphCanvas, LiteGraph } from '@comfyorg/litegraph'
 import { LGraphNode, type NodeId } from '@comfyorg/litegraph/dist/LGraphNode'
 
 import { t } from '@/i18n'
+import {
+  ComfyLink,
+  ComfyNode,
+  ComfyWorkflowJSON
+} from '@/schemas/comfyWorkflow'
 import { useDialogService } from '@/services/dialogService'
 import { useNodeDefStore } from '@/stores/nodeDefStore'
 import { useToastStore } from '@/stores/toastStore'
 import type { ComfyNodeDef } from '@/types/apiTypes'
 import { ComfyExtension } from '@/types/comfy'
-import { ComfyLink, ComfyNode, ComfyWorkflowJSON } from '@/types/comfyWorkflow'
 import { deserialiseAndCreate, serialise } from '@/utils/vintageClipboard'
 
 import { api } from '../../scripts/api'
