+++ tests-ui/tests/comfyWorkflow.test.ts
@@ -1,8 +1,8 @@
 // @ts-strict-ignore
 import fs from 'fs'
 
+import { validateComfyWorkflow } from '@/schemas/comfyWorkflow'
 import { defaultGraph } from '@/scripts/defaultGraph'
-import { validateComfyWorkflow } from '@/types/comfyWorkflow'
 
 const WORKFLOW_DIR = 'tests-ui/workflows'
